move userlogin to another module
